moved core data to separate lib
