+++ CMakeLists.txt
@@ -132,7 +132,7 @@ add_subdirectory(protocol)
 add_subdirectory(primitives)
 add_subdirectory(common)
 add_subdirectory(nrghash)
-#add_subdirectory(miners)
+add_subdirectory(nrgcore)
 
 if (ETHASHCL)
     add_subdirectory(libegihash-cl)
