+++ energiminer/CMakeLists.txt
@@ -16,14 +16,7 @@ add_executable(${EXECUTABLE} ${SRC_LIST} ${SOURCES} ${HEADERS})
 
 add_dependencies(${EXECUTABLE} BuildInfo.h )
 
-target_link_libraries(${EXECUTABLE} libjson-rpc-cpp::client poolprotocols libprimitives libcommon libnrghash Boost::boost Threads::Threads Boost::system)
-
-if (ETHASHCL)
-    target_link_libraries(${EXECUTABLE} egihash-cl)
-endif()
-if (ETHASHCUDA)
-    target_link_libraries(${EXECUTABLE} nrghash-cuda)
-endif()
+target_link_libraries(${EXECUTABLE} libjson-rpc-cpp::client poolprotocols libprimitives libcommon libnrgcore Boost::boost Threads::Threads Boost::system)
 
 include(GNUInstallDirs)
 install(TARGETS energiminer DESTINATION ${CMAKE_INSTALL_BINDIR})
