+++ energiminer/TestMiner.h
@@ -9,8 +9,8 @@
 #define ENERGIMINER_TESTMINER_H_
 
 
-#include "energiminer/plant.h"
-#include "energiminer/miner.h"
+#include "nrgcore/plant.h"
+#include "nrgcore/miner.h"
 
 
 namespace energi
