+++ libegihash-cl/OpenCLMiner.h
@@ -14,8 +14,8 @@
 #define CL_HPP_TARGET_OPENCL_VERSION 120
 #define CL_HPP_MINIMUM_OPENCL_VERSION 120
 
-#include "energiminer/plant.h"
-#include "energiminer/miner.h"
+#include "nrgcore/plant.h"
+#include "nrgcore/miner.h"
 
 #include <cstdint>
 #include <mutex>
