+++ nrgcore/miner.cpp
