+++ primitives/CMakeLists.txt
@@ -1,5 +1,5 @@
 file(GLOB SOURCES "*.cpp")
-file(GLOB SEADERS "*.h")
+file(GLOB HEADERS "*.h")
 
 add_library(libprimitives ${SOURCES} ${HEADERS})
 target_link_libraries(libprimitives PRIVATE jsoncpp_lib_static)
